+++ src/app/features/public/pages/about/about.component.scss
