+++ src/app/features/public/pages/about/components/about-text/about-text.component.spec.ts
